Init Dockerfile & containerized pulpito-ng
Motivation:

We need a way to connect `pulpito-ng`
with `teuthology-api` + all other teuthology services.
Currently, we make use of docker containers to create
a development environment for `teuthology-api` and
other teuthology services.

Solution:

Created a Dockerfile and added a section in the README
file describing the process of connecting `pulpito-ng`
with otheer teuthology services such that we will be able
to containerize `pulpito-ng` and develop it along side
other services.

Signed-off-by: Kamoltat Sirivadhna <ksirivad@redhat.com>