Update DataGrid to v6
This turned into a pretty major change. The breaking changes in the new
DataGrid API necessitated a change in how we were storing and updating
state. The reducers were dropped in favor of the use-query-params
library. As a bonus, filtering, pagination, and deep links all work
better than they did before.

Signed-off-by: Zack Cerza <zack@redhat.com>